+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Unreleased
 
 ### Added
 * Added documentation on the aggregated-translations tool
+* Added services.default-config to quickly reference the default values used by the wdio services
 
 3.0.1 - (April 24, 2018)
 ----------
@@ -23,7 +24,6 @@ Unreleased
 
 ### Added
 * Added nl and nl-BE to supported locales
-
 
 2.11.0 - (April 11, 2018)
 ----------